requirements.txt
PyJWT>=2.8
streamlit-js-eval>=0.1.7
fastapi
httpx
pydantic
pytest
pytest-cov
python-dotenv>=1.0
requests
requests>=2.28
SQLAlchemy==2.0.49
streamlit>=1.32
uvicorn==0.44.0

